WIP: Cleanup realtime assistant merger more
